Runna app/main (#242)
From https://github.com/statsig-io/ios-sdk/pull/11

---------

+++ Sources/Statsig/EvaluationDetails.swift
@@ -4,14 +4,14 @@ public struct EvaluationDetails: Codable {
     let source: EvaluationSource;
 
     var reason: EvaluationReason?;
-    var lcut: UInt?
-    var receivedAt: UInt?
+    var lcut: UInt64?
+    var receivedAt: UInt64?
 
     init(
         source: EvaluationSource,
         reason: EvaluationReason? = nil,
-        lcut: UInt? = nil,
-        receivedAt: UInt? = nil
+        lcut: UInt64? = nil,
+        receivedAt: UInt64? = nil
     ) {
         self.source = source
         self.reason = reason
